# server/doc/swagger.yaml
openapi: "3.0.0"
servers: []
info:
  description: "API for epitech AREA project"
  version: "1.0.0"
  title: Area API
...
paths:
...
  /areas:
    post:
      tags:
        - "area"
      summary: "Create area"
      operationId: "createArea"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                schema:
                  $ref: '#/components/schemas/Area'
      responses:
        '201':
          description: "Area created"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Area'
        '400':
          description: "Invalid input, object invalid"
    get:
      tags:
        - "area"
      summary: Get Aareas
      operationId: "getAreas"
      responses:
        '200':
          description: "Successful operation"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Area'
  /areas/{id}:
    get:
      tags:
        - "area"
      summary: Get area by id
      operationId: "getAreaByIds"
      parameters:
        - in: path
          name: id
          description: Id of the area
          required: true
          schema:
            type: string
      responses:
        '200':
          description: "Successful operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Area'
        '400':
          description: "Invalid id supplied"
        '404':
          description: "Area not found"
    put:
      tags:
        - "area"
      summary: Update service by id
      operationId: "updateArea"
      parameters:
        - in: path
          name: id
          description: Id of the area
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Spotify
                user_id:
                  type: string
                  example: 15796217862479861
                access_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjVjNGEyYjNiZTQ4NzY0N2RlMGFiMTAxZSJ9.inJ1AgLNOxBnv_lizbvdeKsrKszHGctrUsAUdyIZYys
                areas:
                  type: array
                  items:
                    type: string
                    example: "emptyArray"
      responses:
        '200':
          description: "Successful operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Area'
        '400':
          description: "Invalid id supplied"
    delete:
      tags:
        - "area"
      summary: delete user by area
      operationId: "deleteArea"
      parameters:
        - in: path
          name: id
          description: Id of the service
          required: true
          schema:
            type: string
      responses:
        '204':
          description: "Successful operation"
        '400':
          description: "Invalid id supplied"
        '404':
          description: "Area not found"
components:
  schemas:
...
    Area:
      type: object
      properties:
        id:
          type: string
          example: 5c2680ece83bfb2a4c978c9b
        name:
          type: string
          example: SpotifyPlaylist
        params:
          type: array
          items:
            type: string
            example: parameter